+++ bin/uechoctl/uechoctl.rs
@@ -21,6 +21,7 @@ use echonet::Controller;
 use echonet::log::Logger;
 use echonet::protocol::{ESV, Message, Property};
 use echonet::util::Bytes;
+use echonet::{ManufactureCode, StandardDatabase};
 
 #[derive(Parser)]
 #[command(author, version, about = "ECHONET Lite controller utility")]
@@ -34,6 +35,12 @@ struct Cli {
 
 #[derive(Subcommand)]
 enum Commands {
+    #[command(about = "Search ECHONET Lite nodes")]
+    Scan {
+        #[arg(short, long, help = "Request all properties")]
+        all: bool,
+    },
+
     #[command(about = "Send a message to a remote node")]
     Set {
         #[arg(value_name = "IP address")]
@@ -162,6 +169,87 @@ fn run_set(
     Ok(())
 }
 
+fn run_scan(all: bool) -> Result<(), String> {
+    let only_mandatory_properties = !all;
+
+    let mut ctrl = Controller::new();
+    ctrl.start();
+    ctrl.search();
+
+    thread::sleep(time::Duration::from_secs(2));
+
+    for (i, node) in ctrl.nodes().iter().enumerate() {
+        // Makes a manufacture code read (ESV::ReadRequest) message.
+        let mut msg = Message::new();
+        msg.set_esv(ESV::ReadRequest);
+        msg.set_deoj(0x0EF001);
+        let mut prop = Property::new();
+        prop.set_code(0x8A);
+        msg.add_property(prop);
+
+        let mut manufacture_name = String::from("");
+        let rx = ctrl.post_message(&node, &mut msg);
+        match rx.recv_timeout(Duration::from_secs(1)) {
+            Ok(msg) => {
+                let props = msg.properties();
+                if 0 < props.len() {
+                    let manufacture_code = Bytes::to_u32(props[0].data()) as ManufactureCode;
+                    let std_db = StandardDatabase::shared();
+                    let found_manufacture = std_db.find_manufacture(manufacture_code);
+                    if found_manufacture.is_some() {
+                        manufacture_name = found_manufacture.unwrap().name().to_string();
+                    }
+                }
+            }
+            Err(_e) => {}
+        };
+
+        println!("[{}] {} ({})", i, node.addr(), manufacture_name);
+
+        // Prints all mandatory properties in the object.
+
+        for (j, obj) in node.objects().iter().enumerate() {
+            println!("    [{}] {:06X} ({})", j, obj.code(), obj.class_name());
+            for obj_prop in obj.properties() {
+                if only_mandatory_properties && !obj_prop.is_read_required() {
+                    continue;
+                }
+
+                // Makes a property value read (ESV::ReadRequest) message.
+                let mut msg = Message::new();
+                msg.set_esv(ESV::ReadRequest);
+                msg.set_deoj(obj.code());
+                let mut prop = Property::new();
+                prop.set_code(obj_prop.code());
+                msg.add_property(prop);
+
+                let mut prop_data = String::from("");
+                let rx = ctrl.post_message(&node, &mut msg);
+                match rx.recv_timeout(Duration::from_secs(1)) {
+                    Ok(msg) => {
+                        if 0 < msg.opc() {
+                            prop_data = hex::encode(msg.property(0).data());
+                        }
+                    }
+                    Err(e) => {
+                        prop_data = format!("{}", e);
+                    }
+                };
+                println!(
+                    "        [{:02X}] {}: {}",
+                    obj_prop.code(),
+                    obj_prop.name(),
+                    prop_data
+                );
+            }
+        }
+    }
+
+    ctrl.stop();
+
+    Ok(())
+}
+
 fn main() -> Result<(), String> {
     let cli = Cli::parse();
 
@@ -170,6 +258,7 @@ fn main() -> Result<(), String> {
     }
 
     match cli.command {
+        Commands::Scan { all } => run_scan(all),
         Commands::Set {
             node_addr,
             object_code,